Fixed spam of svn message on startup Fixed compilation warning
git-svn-id: https://codelite.svn.sourceforge.net/svnroot/codelite/trunk@5727 9da81c78-c036-0410-9e1f-a2b0375e4b5a

+++ Outline/outline.cpp
@@ -107,7 +107,7 @@ SymbolViewPlugin::SymbolViewPlugin(IManager *manager)
 			
         else {
 			size_t count = book->GetPageCount();
-			if(index >= count) {
+			if(index >= (int)count) {
 				// Invalid index
 				book->AddPage(m_view, wxT("Outline"), false);
